reviewer from hpe added for repo maintenance
Opening discussions re: phosphor-u-boot-env-mgr maintenance.

HPE is working on a service that reads HW ID data (to be consumed by
Entity-Manager probes) from device tree nodes that are modified by code
in u-boot. After a number of design doc adjustments, we seem to be
nearing a community consensus re: a design that has the daemon hosted
in phosphor-u-boot-env-mgr.

https://gerrit.openbmc.org/c/openbmc/docs/+/66369
('phosphor-u-boot-env-mgr' is mentioned by name several times
throughout the latest revision)

It's been strongly suggested (charge led by Ed Tanous) we look at
hosting this service in phosphor-u-boot-env-mgr. It's also been
strongly suggested that if HPE is going to leverage this repo,
then we should offer to take over some of the maintenance for it.

Chris Sides is offering to take on a 'reviewer' role to start,
with an eye on eventually becoming a maintainer. If needed in the
future, other (well-experienced with OBMC) HPE reviewers and/or
maintainers can be added.

Change-Id: Ic3ca07cca48a34487d37514aaa973cbeff0847f1
Signed-off-by: Chris Sides <christopher.sides@hpe.com>