switching from cmakeLists to meson build scripts
Nearly all OBMC subprojects are using MESON. This change brings this
repo in line with other active OBMC repos, in addition to clarifying
the build process for easier future additions via simpler build
configuration scripting requirements in MESON.

The MESON scripts used here are primed to allow multiple separate
'features' to be enabled/disabled at build time via the recipe.
For now, the only 'feature' is the u-boot-env-mgr daemon, but
(at least one) more services are expected to be added in the future.

Change-Id: Ie3c7a4a49123e49cca652151bfdfc4e67bd0a330
Signed-off-by: Chris Sides <Christopher.Sides@hpe.com>