service: device-tree node to d-bus for hw id data
A daemon for transferring Viable Product Data (VPD) from device-tree
nodes to D-Bus as described in the design doc @
https://gerrit.openbmc.org/c/openbmc/docs/+/66369

This checkin is not intended for +2 acceptance in the
phosphor-u-boot-env-mgr repo,  but for discussion of
code that will be pushed to a yet-to-be created
devicetree-vpd-parser repo

Tested:
Ensure 'model' and/or 'serial-number' nodes are populated in
/proc/device-tree. Can be hardcoded into platform DTS via
linux-kernel recipe if needed.

'''
After OBMC boot, calling
> busctl introspect xyz.openbmc_project.Inventory.MachineContext /xyz/openbmc_project/Inventory/MachineContext

produces the following output:

NAME                                          TYPE      SIGNATURE  RESULT/VALUE  FLAGS
...
xyz.openbmc_project.Inventory.Decorator.Asset interface -          -             -
.BuildDate                                    property  s          ""            emits-change writable
.Manufacturer                                 property  s          ""            emits-change writable
.Model                                        property  s          "hpe,dl360"   emits-change writable
.PartNumber                                   property  s          ""            emits-change writable
.SerialNumber                                 property  s          "t5texpl"     emits-change writable
...

Note: Writing to one of these properties will update the property on D-Bus without affecting the
underlying device-tree node. An 'updated' event will be fired, so Entity-Manager probes monitoring
a given property would recognize the new value and key off it.
'''

Change-Id: I9e8d688284c8a8aedba0b07ecfe0faaf313bc015
Signed-off-by: Chris Sides <Christopher.Sides@hpe.com>